requirements.txt
# Marketing Audit System Dependencies
# Install with: pip install -r requirements.txt

# Anthropic Claude API SDK
anthropic>=0.39.0

# Web scraping and content extraction
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment variable management
python-dotenv>=1.0.0

# HTML parsing
lxml>=5.0.0

# Optional: For enhanced error handling and retries
urllib3>=2.0.0
